+++ src/descriptor.rs
@@ -178,6 +178,7 @@ impl DescriptorWriter<'_> {
     ///   that do not conform to any class.
     /// * `function_sub_class` - Sub-class code. Depends on class.
     /// * `function_protocol` - Protocol code. Depends on class and sub-class.
+    /// * `function_string` - Index of string descriptor describing this function
     pub fn iad(
         &mut self,
         first_interface: InterfaceNumber,
@@ -185,10 +186,13 @@ impl DescriptorWriter<'_> {
         function_class: u8,
         function_sub_class: u8,
         function_protocol: u8,
+        function_string: Option<StringIndex>,
     ) -> Result<()> {
         if !self.write_iads {
             return Ok(());
         }
+
+        let str_index = function_string.map_or(0, Into::into);
 
         self.write(
             descriptor_type::IAD,
@@ -198,7 +202,7 @@ impl DescriptorWriter<'_> {
                 function_class,
                 function_sub_class,
                 function_protocol,
-                0,
+                str_index,
             ],
         )?;